Started work on implementing Cards. Values can add, subtract, and iterate.
Implemented Add and Sub for Value with i32. Negative cases for Add should
be handled with Sub and vice versa. Maybe in the future I can switch back to
u32 for Value::Number, we'll see.

Created an iterator for Value as a struct that just has a counter. I'm sure
there is a much more elegant way to do this going down the line, just not
quite sure what that is right now.